refactor: split packaging_agent.py and packaging_validator.py into packaging/ subpackage (#1278)
- Extract validator.py with ValidationSeverity, ValidationIssue, ValidationResult, PackagingValidator
- Extract manifest.py with ManifestGenerator for manifest.json and pack_icon generation
- Extract folder_builder.py for behaviors/resources/scripts folder structure
- Extract bundler.py for dependency resolution and MCADDON archive creation
- Extract pack_report.py for packaging validation report assembly
- Add __init__.py with PackagingCoordinator thin wrapper
- Update packaging_agent.py to delegate to packaging/ subpackage
- Update packaging_validator.py to re-export from packaging/ subpackage
- All 52 packaging tests pass, public API unchanged
